Move processing in main flow
