+++ client/landing/stepper/declarative-flow/videopress.ts
@@ -1,9 +1,14 @@
+import { PlansSelect, SiteSelect } from '@automattic/data-stores';
 import { useLocale } from '@automattic/i18n-utils';
 import { useFlowProgress, VIDEOPRESS_FLOW } from '@automattic/onboarding';
 import { useSelect, useDispatch } from '@wordpress/data';
 import { translate } from 'i18n-calypso';
+import { useSupportedPlans } from 'calypso/../packages/plans-grid/src/hooks';
+import { useNewSiteVisibility } from 'calypso/landing/stepper/hooks/use-selected-plan';
+import { domainRegistration } from 'calypso/lib/cart-values/cart-items';
+import { cartManagerClient } from 'calypso/my-sites/checkout/cart-manager-client';
 import { useSiteSlug } from '../hooks/use-site-slug';
-import { USER_STORE, ONBOARD_STORE } from '../stores';
+import { PLANS_STORE, SITE_STORE, USER_STORE, ONBOARD_STORE } from '../stores';
 import './internals/videopress.scss';
 import ChooseADomain from './internals/steps-repository/choose-a-domain';
 import Intro from './internals/steps-repository/intro';
@@ -65,6 +70,29 @@ const videopress: Flow = {
 		);
 		const locale = useLocale();
 
+		const { createVideoPressSite, setSelectedSite, setPendingAction, setProgress } =
+			useDispatch( ONBOARD_STORE );
+		const currentUser = useSelect(
+			( select ) => ( select( USER_STORE ) as UserSelect ).getCurrentUser(),
+			[]
+		);
+		const siteDescription = useSelect(
+			( select ) => ( select( ONBOARD_STORE ) as OnboardSelect ).getSelectedSiteDescription(),
+			[]
+		);
+		const getPlanProduct = useSelect(
+			( select ) => ( select( PLANS_STORE ) as PlansSelect ).getPlanProduct,
+			[]
+		);
+		const visibility = useNewSiteVisibility();
+		const { getNewSite } = useSelect( ( select ) => select( SITE_STORE ) as SiteSelect, [] );
+		const { saveSiteSettings, setIntentOnSite } = useDispatch( SITE_STORE );
+		const { supportedPlans } = useSupportedPlans( locale, 'MONTHLY' );
+		const selectedDomain = useSelect(
+			( select ) => ( select( ONBOARD_STORE ) as OnboardSelect ).getSelectedDomain(),
+			[]
+		);
+
 		const clearOnboardingSiteOptions = () => {
 			setSiteTitle( '' );
 			setSiteDescription( '' );
@@ -93,6 +121,77 @@ const videopress: Flow = {
 			return true;
 		};
 
+		const addVideoPressPendingAction = () => {
+			setPendingAction( async () => {
+				setProgress( 0 );
+				try {
+					await createVideoPressSite( {
+						username: currentUser ? currentUser?.username : '',
+						languageSlug: locale,
+						visibility,
+					} );
+				} catch ( e ) {
+					return;
+				}
+				setProgress( 0.5 );
+
+				const newSite = getNewSite();
+				setSelectedSite( newSite?.blogid );
+				setIntentOnSite( newSite?.site_slug as string, VIDEOPRESS_FLOW as string );
+				saveSiteSettings( newSite?.blogid as number, {
+					launchpad_screen: 'full',
+					blogdescription: siteDescription,
+				} );
+
+				let planObject = supportedPlans.find( ( plan ) => 'premium' === plan.periodAgnosticSlug );
+				if ( ! planObject ) {
+					planObject = supportedPlans.find( ( plan ) => 'business' === plan.periodAgnosticSlug );
+				}
+
+				const planProductObject = getPlanProduct( planObject?.periodAgnosticSlug, 'MONTHLY' );
+
+				const cartKey = await cartManagerClient.getCartKeyForSiteSlug(
+					newSite?.site_slug as string
+				);
+
+				// eslint-disable-next-line @typescript-eslint/no-explicit-any
+				const productsToAdd: any[] = [
+					{
+						product_slug: planProductObject?.storeSlug,
+						extra: {
+							signup_flow: VIDEOPRESS_FLOW,
+						},
+					},
+				];
+
+				if ( selectedDomain && selectedDomain.product_slug ) {
+					const registration = domainRegistration( {
+						domain: selectedDomain.domain_name,
+						productSlug: selectedDomain.product_slug as string,
+						extra: { privacy_available: selectedDomain.supports_privacy },
+					} );
+
+					productsToAdd.push( registration );
+				}
+
+				setProgress( 0.75 );
+
+				cartManagerClient
+					.forCartKey( cartKey )
+					.actions.addProductsToCart( productsToAdd )
+					.then( () => {
+						setProgress( 1.0 );
+						const redirectTo = encodeURIComponent(
+							`/setup/videopress/launchpad?siteSlug=${ newSite?.site_slug }&siteId=${ newSite?.blogid }`
+						);
+
+						window.location.replace(
+							`/checkout/${ newSite?.site_slug }?signup=1&redirect_to=${ redirectTo }`
+						);
+					} );
+			} );
+		};
+
 		switch ( _currentStep ) {
 			case 'intro':
 				clearOnboardingSiteOptions();
@@ -102,6 +201,9 @@ const videopress: Flow = {
 				break;
 			case 'chooseADomain':
 				stepValidateSiteTitle();
+				break;
+			case 'processing':
+				addVideoPressPendingAction();
 				break;
 		}
